refactor: useChapterDomain에서 목업 타입 의존성 제거
